+++ config.py
@@ -1,0 +1,13 @@
+DB_PATH = "users.db"
+ADMIN_ROLE_ID = 1351479405699928108
+SHEHER_ROLE_ID = 1351546875492302859
+HEHIM_ROLE_ID = 1351546766855634984
+MAX_COINS = 9223372036854775807
+DAILY_REWARD = 20
+WELCOME_CHANNEL_ID = 1351487186557734942
+LOG_CHANNEL_ID = 1364226902289813514
+STAT_PRICE = 66
+QUEST_COOLDOWN_HOURS = 3
+FISHING_COOLDOWN_MINUTES = 30
+WEEKLY_REWARD = 50
+STAT_NAMES = ["intelligence", "strength", "stealth"]
